Add command to auto-assign participants to their computers
Automatically assigns participants to computers based on a JSON file. It reads participant-computer mappings from the provided file and updates the database by linking each participant to the corresponding computer based on either IP address or a predefined name.

The JSON file should be a list of pairs, where each entry is formatted as `[participant_id, computer_identifier]`. The `computer_identifier` can be either an IP address or a predefined name used to identify the computer in the database."